+++ src/Engine/Vector.cpp
@@ -1,7 +1,7 @@
 #include "Vector.h"
 #include <math.h>
 
-Vector::Vector(float x, float y, float z)
+Vector::Vector(double x, double y, double z)
 {
     this->x = x;
     this->y = y;
@@ -15,17 +15,17 @@ Vector::Vector(const Point &start, const Point &end)
     z = end.z - start.z;
 }
 
-float Vector::length() const
+double Vector::length() const
 { 
     return sqrt(x * x + y * y + z * z); 
 }
 
-float Vector::sqrLength() const
+double Vector::sqrLength() const
 { 
     return x * x + y * y + z * z; 
 }
 
-float & Vector::operator[](int index)
+double & Vector::operator[](int index)
 {
     if (index == 0)
         return x;
@@ -35,7 +35,7 @@ float & Vector::operator[](int index)
         return z;
 }
 
-const float & Vector::operator[](int index) const
+const double & Vector::operator[](int index) const
 {
     if (index == 0)
         return x;
@@ -55,7 +55,7 @@ Vector Vector::operator-(const Vector &b) const
     return Vector(x - b.x, y - b.y, z - b.z); 
 }
 
-Vector Vector::operator*(float b) const 
+Vector Vector::operator*(double b) const 
 { 
     return Vector(x * b, y * b, z * b); 
 }
@@ -70,12 +70,12 @@ Vector& Vector::norm()
     return *this = *this * (1 / sqrt(x * x + y * y + z * z)); 
 }
 
-float Vector::dot(const Vector &b) const 
+double Vector::dot(const Vector &b) const 
 {
     return x * b.x + y * b.y + z * b.z; 
 }
 
-float Vector::dot(const Point &p) const
+double Vector::dot(const Point &p) const
 {
     return x * p.x + y * p.y + z * p.z;
 }
@@ -85,7 +85,7 @@ Vector Vector::cross(const Vector &b)
     return Vector(y * b.z - z * b.y, z * b.x - x * b.z, x * b.y - y * b.x);
 }
 
-float Vector::angleTo(const Vector &b) const
+double Vector::angleTo(const Vector &b) const
 {
     return acos(dot(b) * (1.0f / (length() * b.length())));
 }
